iam/requirements.txt
fastapi==0.128.0
beanie==1.26.0
motor==3.4.0
pymongo==4.6.3
pydantic==2.12.5
pydantic[email]
pydantic_settings==2.12.0
python-dotenv==1.0.1
uvicorn==0.32.0
httpx==0.28.1
aiohttp==3.13.3
auth-gateway-serverkit==0.0.82
